Notify bisection infra when testcase is marked non-security.
Also move the bisection module from bot.tasks to base for use from App
Engine.
